[1824] Refactor to put all Cisleithania specifics in g_1824_cisleithania
