+++ lib/engine/game/g_1824/corporation.rb
@@ -27,9 +27,6 @@ module Engine
           end
           @real_presidents_percent = @presidents_share.percent
 
-          # Used for 2 player variant initial SR
-          @stack = nil
-
           # Used for coal railway valuation
           @coal_price = 0
         end
@@ -94,28 +91,6 @@ module Engine
           @floatable = true
           @percent_total_ipo_shares = 100
           @real_presidents_percent = @presidents_share.percent
-        end
-
-        def make_construction_railway!
-          @type = :construction_railway
-          a = @abilities.first
-          remove_ability(a)
-        end
-
-        def make_bond_railway!
-          @type = :bond_railway
-          remove_reserve_for_all_shares!
-          @ipoed = true
-          float!
-
-          # Presidency share is treated as a double cert
-          @presidents_share.double_cert = true
-        end
-
-        def receivership?
-          return true if @type == :bond_railway
-
-          super
         end
       end
     end
