+++ lib/engine/game/g_1824/map.rb
@@ -8,8 +8,8 @@ module Engine
 
         AXES = { x: :number, y: :letter }.freeze
 
-        # Used by 1837, not used in 1824
-        MINE_HEXES = [].freeze
+        # Standard game has 4 mine hexes
+        MINE_HEXES = %w[C6 A12 A22 H25].freeze
 
         # Used by 1837, not used in 1824
         ITALY_HEXES = [].freeze
@@ -230,7 +230,7 @@ module Engine
         PLAIN_WITH_MOUNTAIN = 'upgrade=cost:40,terrain:mountain'
         PLAIN_WITH_WATER = 'upgrade=cost:20,terrain:water'
 
-        def base_map
+        def map_optional_hexes
           plain_hexes = %w[B7 B11 B17 B19 B21 C8 C14 C20 C22 C24 D9 D11 D13 D15 D17 E6 E18 E22
                            F9 F13 F15 F21 F25 G6 G12 G14 G22 G24 H9 H13 H19 H21 I10 I12 I14]
           one_town = %w[A8 A20 C10 C16 D25 E20 E24 F19 G2 G20 H11 I20 I22]
@@ -283,92 +283,6 @@ module Engine
               ['E16'] => DOUBLE_TOWN_WITH_WATER,
             },
           }
-        end
-
-        def cisleithania_map
-          # For 3 players Budapest is a city for Pre-staatsbahn U1
-          budapest = @players.size == 3 ? 'city' : 'offboard'
-          b_major = ';icon=image:1824/B,sticky:1'
-          b_minor = ';icon=image:1824/b_lower_case,sticky:1'
-          {
-            red: {
-              ['A4'] => DRESDEN_1,
-              ['A24'] => KIEW_1,
-              ['A26'] => KIEW_2,
-              ['B3'] => DRESDEN_2,
-              ['E14'] =>
-                'offboard=revenue:yellow_20|green_30|brown_40|gray_40;path=a:0,b:_0,terminal:1;'\
-                'path=a:1,b:_0,terminal:1;path=a:2,b:_0,terminal:1;path=a:3,b:_0,terminal:1',
-              ['G12'] =>
-                "#{budapest}=revenue:yellow_20|green_40|brown_60|gray_70;"\
-                'path=a:0,b:_0,terminal:1;path=a:1,b:_0,terminal:1;path=a:2,b:_0,terminal:1;path=a:3,b:_0,terminal:1',
-              ['F25'] =>
-                'offboard=revenue:yellow_20|green_30|brown_40|gray_40;path=a:2,b:_0,terminal:1;'\
-                'path=a:3,b:_0,terminal:1',
-              ['H1'] => MAINLAND_1,
-              ['I2'] => MAINLAND_2,
-              ['I10'] =>
-                'offboard=revenue:yellow_10|green_10|brown_50|gray_50;path=a:1,b:_0,terminal:1;'\
-                'path=a:2,b:_0,terminal:1;path=a:3,b:_0,terminal:1',
-            },
-            gray: {
-              ['A12'] => MINE_2,
-              ['A22'] => MINE_3,
-              ['B17'] => 'path=a:0,b:3;path=a:1,b:4;path=a:1,b:3;path=a:0,b:4',
-              ['C6'] => MINE_1,
-            },
-            white: {
-              %w[A8 A20 C10 C16 G2] => TOWN,
-              %w[D25 E24] => "#{TOWN}#{b_minor}",
-              %w[A6 A10] => TOWN_WITH_MOUNTAIN,
-              %w[B13 B25 F11] => DOUBLE_TOWN_WITH_WATER,
-              %w[H3] => CITY_WITH_MOUNTAIN,
-              %w[A18 C26 I8] => CITY_LABEL_T,
-              ['E26'] => "#{CITY_LABEL_T}#{b_minor}",
-              %w[B5 B15 B23 C12 E8 F7 G4 G10] => CITY,
-              ['B9'] => "#{CITY}#{b_major}",
-              %w[B7 B11 B19 B21 C8 C14 C20 C22 C24 D9 D11 D13 D15 E6
-                 F9 F13 G6 H9 H11] => PLAIN,
-              %w[D23 G8 H5 H7] => PLAIN_WITH_MOUNTAIN,
-              %w[E10] => PLAIN_WITH_WATER,
-              ['E12'] => "#{WIEN}#{b_major}",
-            },
-          }
-        end
-
-        def show_map_legend?
-          option_cisleithania
-        end
-
-        def map_legends
-          ['bonus_legend']
-        end
-
-        def bonus_legend(_font_color, *_extra_colors)
-          [
-            # table-wide props
-            {
-              style: {
-                margin: '0.5rem 0 0.5rem 0',
-                border: '1px solid',
-                borderCollapse: 'collapse',
-              },
-            },
-            # header
-            [
-              { text: 'Bonus', props: { style: { border: '1px solid' } } },
-              { text: 'Run', props: { style: { border: '1px solid' } } },
-              { text: 'Icons', props: { style: { border: '1px solid' } } },
-              { text: 'G', props: { style: { border: '1px solid' } } },
-            ],
-            # body
-            [
-              { text: 'Bukowina', props: { style: { border: '1px solid' } } },
-              { text: 'Prag/Wien - Bukowina', props: { style: { textAlign: 'center', border: '1px solid' } } },
-              { text: 'B-b', props: { style: { textAlign: 'center', border: '1px solid' } } },
-              { text: '50', props: { style: { textAlign: 'right', border: '1px solid' } } },
-            ],
-          ]
         end
       end
     end
