+++ lib/engine/game/g_1824/step/buy_sell_par_exchange_shares.rb
@@ -47,8 +47,7 @@ module Engine
           def can_sell?(_entity, bundle)
             # Rule VI.8, bullet 1, sub-bullet 2: Bank ownership cannot exceed 50% for started corporations
             corp = bundle.corporation
-            super && @game.buyable?(corp) &&
-                    (@game.bond_railway?(corp) || (corp.ipo_shares.sum(&:percent) + bundle.percent <= 50))
+            super && @game.buyable?(corp) && (corp.ipo_shares.sum(&:percent) + bundle.percent <= 50)
           end
 
           # Rule VI.7, bullet 4: Exchange can take you over 60%
@@ -73,11 +72,9 @@ module Engine
               (@game.num_certs(entity) < @game.cert_limit(entity)) && @game.buyable?(corporation)
           end
 
-          # Rule X.4, bullet 2: Maybe exceed 60% in 2 player 1824, if buying from market
-          def allowed_buy_from_market(_entity, bundle)
-            return false unless @game.two_player?
-
-            bundle.shares.first.owner == @game.share_pool
+          # Needed for two player variant, see Cisleithania implementation
+          def allowed_buy_from_market(_entity, _bundle)
+            false
           end
 
           def process_buy_shares(action)
