+++ lib/engine/game/g_1824/step/buy_train.rb
@@ -15,9 +15,6 @@ module Engine
             # Rule VII.11 all entities must own a train, unless coal company and there are no g-trains in the depot
             return false unless entity.trains.empty?
 
-            # Rule X.3 and X.4: Construction railways cannot own any trains
-            return false if @game.construction_railway?(entity) || @game.bond_railway?(entity)
-
             return true unless @game.coal_railway?(entity)
 
             depot_g_trains = @depot.depot_trains.select { |t| @game.goods_train?(t.name) }
@@ -49,12 +46,6 @@ module Engine
             super
 
             @game.two_train_bought = true if train.name == '2'
-
-            return unless @game.two_player?
-
-            # Rule X.4, need to handle extra tokening of bond railway
-            @game.set_last_train_buyer(entity, train) if train.name == '4' && @depot.depot_trains.first.name == '5'
-            @game.set_last_train_buyer(entity, train) if train.name == '5' && @depot.depot_trains.first.name == '6'
           end
 
           def buyable_trains(entity)
