+++ lib/engine/game/g_1824/step/kk_token_choice.rb
@@ -55,7 +55,7 @@ module Engine
 
             choice = action.choice == 'Return one token' ? 1 : 2
             @log << "#{player.name} removes selected #{@game.kk.name} token from Wien, and put it as a "\
-                    "#{@game.format_currency(100)} token on the #{@game.kk.name} charter"
+                    "#{@game.format_currency(40)} token on the #{@game.kk.name} charter"
             @game.return_kk_token(choice)
             @game.kk_token_choice_player = nil
           end
