+++ lib/engine/game/g_1824_cisleithania.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Engine
+  module Game
+    module G1824Cisleithania
+    end
+  end
+end
